Split Colonini parser's ValueExpr into new files
