+++ src/bh/config/colonini/internal/parser.h
@@ -27,6 +27,7 @@
 #include "bh/config/colonini/internal/lexer.h"
 #include "bh/config/colonini/internal/parser_type/const_expr.h"
 #include "bh/config/colonini/internal/parser_type/toggle_expr.h"
+#include "bh/config/colonini/internal/parser_type/value_expr.h"
 #include "bh/config/colonini/type.h"
 
 #ifdef __cplusplus
@@ -41,22 +42,6 @@ struct KeyExpr {
   struct ConstExpr constexpr;
   struct Subscript* subscripts;
   size_t subscripts_count;
-};
-
-enum ValueExprType {
-  ValueExprType_kUnspecified,
-
-  ValueExprType_kEmpty,
-  ValueExprType_kConst,
-  ValueExprType_kToggle,
-};
-
-struct ValueExpr {
-  enum ValueExprType type;
-  union {
-    struct ConstExpr as_constexpr;
-    struct ToggleExpr as_toggleexpr;
-  } variant;
 };
 
 struct AssignStatement {
