+++ src/bh/config/colonini/internal/parser_type/CMakeLists.txt
@@ -19,12 +19,18 @@
 
 target_sources(${PROJECT_NAME}
     PRIVATE
-        "${CMAKE_CURRENT_SOURCE_DIR}/const_expr_type.c"
-        "${CMAKE_CURRENT_SOURCE_DIR}/const_expr_type.h"
-
         "${CMAKE_CURRENT_SOURCE_DIR}/const_expr.c"
         "${CMAKE_CURRENT_SOURCE_DIR}/const_expr.h"
 
+        "${CMAKE_CURRENT_SOURCE_DIR}/const_expr_type.c"
+        "${CMAKE_CURRENT_SOURCE_DIR}/const_expr_type.h"
+
         "${CMAKE_CURRENT_SOURCE_DIR}/toggle_expr.c"
         "${CMAKE_CURRENT_SOURCE_DIR}/toggle_expr.h"
+
+        "${CMAKE_CURRENT_SOURCE_DIR}/value_expr.c"
+        "${CMAKE_CURRENT_SOURCE_DIR}/value_expr.h"
+
+        "${CMAKE_CURRENT_SOURCE_DIR}/value_expr_type.c"
+        "${CMAKE_CURRENT_SOURCE_DIR}/value_expr_type.h"
 )
